fix: CDMA lat/lon type mismatch + add gradle wrapper jar
- Fixed CellInfoCollector CDMA latitude/longitude Int→Double conversion
- Generated gradle-wrapper.jar (was missing from initial commit)
- Build verified: assembleDebug passes, 18MB APK produced

+++ app/src/main/java/com/bp22intel/edgesentinel/data/sensor/CellInfoCollector.kt
@@ -130,8 +130,8 @@ class CellInfoCollector @Inject constructor(
                         mnc = identity.systemId,
                         signalStrength = signal.dbm,
                         networkType = NetworkType.CDMA,
-                        latitude = identity.latitude,
-                        longitude = identity.longitude,
+                        latitude = identity.latitude.toDouble(),
+                        longitude = identity.longitude.toDouble(),
                         firstSeen = now,
                         lastSeen = now,
                         timesSeen = 1
